Move ClassLoader handling to ClassLoading
